+++ src/components/index.js
@@ -1,4 +1,5 @@
 // Keep in alphabetical order
+export { default as ActiveColorIndicator } from './ActiveColorIndicator/ActiveColorIndicator';
 export { default as Board } from './Board/Board';
 export { default as Game } from './Game/Game';
 export { default as Piece } from './Piece/Piece';