+++ src/containers/App/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import { Game } from './containers';
+import { Game } from '../.';
 
 const Wrapper = styled.div`
   background-color: lightblue;